Move helpers and weaponMount to modules
